DoodlePath支持旋转。可旋转item可设置围绕某个点旋转

+++ doodle/src/main/java/cn/hzw/doodle/util/DrawUtil.java
@@ -145,7 +145,7 @@ public class DrawUtil {
     }
 
     // 顺时针旋转
-    public static PointF rotatePoint(PointF coords, int degree, float x, float y, float px, float py) {
+    public static PointF rotatePoint(PointF coords, float degree, float x, float y, float px, float py) {
         if (degree % 360 == 0) {
             coords.x = x;
             coords.y = y;
